update(*): try ti fix add comments

+++ src/presenter/popup.js
@@ -1,6 +1,6 @@
-import { UpdateType, CommentsTitle } from '../const.js';
+import { UpdateType, CommentsTitle, UserAction } from '../const.js';
 import { getCurrentDate } from '../utils/date.js';
-import { isEsc, isEnter } from '../utils/common.js';
+import { isEsc, isEnter, isOnline } from '../utils/common.js';
 import { render, rerender, remove } from '../utils/render.js';
 import FilmDetailsView from '../view/film-details.js';
 import FilmDetailsBottomView from '../view/film-details-bottom.js';
@@ -9,6 +9,8 @@ import CommentsTitleView from '../view/comments-title.js';
 import CommentsListView from '../view/comments-list.js';
 import CommentView from '../view/comment.js';
 import NewCommentView from '../view/new-comment.js';
+import CommentModel from '../model/comment.js';
+import { alert } from '../utils/alert.js';
 
 export default class FilmDetailsPresenter {
   constructor(filmDetailsContainer, filmsModel, changeFilm, hideFilmDetails, api) {
@@ -18,8 +20,8 @@ export default class FilmDetailsPresenter {
     this._hideFilmDetails = hideFilmDetails;
     this._api = api;
     this._film = null;
-    this._filmComments = [];
     this._isLoading = true;
+    this._commentsModel = new CommentModel();
     this._isError = false;
     this._prevScrollTop = 0;
     this._filmDetailsView = null;
@@ -36,8 +38,10 @@ export default class FilmDetailsPresenter {
     this._handleAddFavoriteButtonClick = this._handleAddFavoriteButtonClick.bind(this);
     this._handleFormSubmit = this._handleFormSubmit.bind(this);
     this._handleDeleteButtonClick = this._handleDeleteButtonClick.bind(this);
-    this._handleModelEvent = this._handleModelEvent.bind(this);
-    this._filmsModel.addObserver(this._handleModelEvent);
+    this._handleFilmsModelEvent = this._handleFilmsModelEvent.bind(this);
+    this._handleCommentsModelEvent = this._handleCommentsModelEvent.bind(this);
+    this._filmsModel.addObserver(this._handleFilmsModelEvent);
+    this._commentsModel.addObserver(this._handleCommentsModelEvent);
   }
 
   init(film, { loadComments = true } = {}) {
@@ -63,6 +67,10 @@ export default class FilmDetailsPresenter {
     throw new Error('Presenter not initialized');
   }
 
+  get _isSuccess() {
+    return !this._isLoading && !this._isError;
+  }
+
   _handleCloseButtonClick() {
     this._hideFilmDetails();
   }
@@ -80,7 +88,7 @@ export default class FilmDetailsPresenter {
   }
 
   async _handleAddToWatchButtonClick() {
-    let updatedFilm = {
+    const updatedFilm = {
       ...this._film,
       userDetails: {
         ...this._film.userDetails,
@@ -88,13 +96,11 @@ export default class FilmDetailsPresenter {
       },
     };
 
-    updatedFilm = await this._api.updateFilm(updatedFilm);
-
-    this._changeFilm(UpdateType.MINOR, updatedFilm);
+    this._changeFilm(UserAction.UPDATE_USER_DETAILS, UpdateType.MINOR, updatedFilm);
   }
 
   async _handleAddWatchedButtonClick() {
-    let updatedFilm ={
+    const updatedFilm ={
       ...this._film,
       userDetails: {
         ...this._film.userDetails,
@@ -103,13 +109,11 @@ export default class FilmDetailsPresenter {
       },
     };
 
-    updatedFilm = await this._api.updateFilm(updatedFilm);
-
-    this._changeFilm(UpdateType.MINOR, updatedFilm);
+    this._changeFilm(UserAction.UPDATE_USER_DETAILS, UpdateType.MINOR, updatedFilm);
   }
 
   async _handleAddFavoriteButtonClick() {
-    let updatedFilm ={
+    const updatedFilm ={
       ...this._film,
       userDetails: {
         ...this._film.userDetails,
@@ -117,12 +121,14 @@ export default class FilmDetailsPresenter {
       },
     };
 
-    updatedFilm = await this._api.updateFilm(updatedFilm);
-
-    this._changeFilm(UpdateType.MINOR, updatedFilm);
+    this._changeFilm(UserAction.UPDATE_USER_DETAILS, UpdateType.MINOR, updatedFilm);
   }
 
   async _handleDeleteButtonClick(commentId) {
+    if (!isOnline()) {
+      alert('You can not delete comment couse you are offline now');
+      return;
+    }
     try {
       this._commentView.get(commentId).setDeletingStatus();
       await this._api.deleteComment(commentId);
@@ -131,8 +137,9 @@ export default class FilmDetailsPresenter {
         ...this._film,
         comments: this._film.comments.filter((id) => id !== commentId),
       };
-      this._filmComments = this._filmComments.filter(({ id }) => id !== commentId);
-      this._changeFilm(UpdateType.PATCH, updatedFilm);
+      await this._api.updateFilm(updatedFilm, { isServerUpdate: false });
+      this._changeFilm(UserAction.DELETE_COMMENT, UpdateType.PATCH, updatedFilm);
+      this._commentsModel.deleteComment(UpdateType.PATCH, commentId);
 
     } catch (error) {
       this._commentView.get(commentId).resetDeletingStatus();
@@ -142,14 +149,19 @@ export default class FilmDetailsPresenter {
 
 
   async _handleFormSubmit() {
+    if (!isOnline()) {
+      this._newCommentView.setErrorState();
+      alert('You can not delete comment couse you are offline now');
+      return;
+    }
     const newComment = this._newCommentView.getData();
     try {
       this._newCommentView.disable();
       this._newCommentView.clearErrorState();
       const { updatedFilm, updatedComments } = await this._api.addComment(this._film.id, newComment);
 
-      this._filmComments = updatedComments;
-      this._changeFilm(UpdateType.PATCH, updatedFilm);
+      this._changeFilm(UserAction.CREATE_COMMENT, UpdateType.PATCH, updatedFilm);
+      this._commentsModel.setComments(UpdateType.PATCH, updatedComments);
 
       this._newCommentView.reset();
 
@@ -159,12 +171,34 @@ export default class FilmDetailsPresenter {
     this._newCommentView.enable();
   }
 
-  _handleModelEvent(updateType, updatedFilm) {
-    if (updateType === UpdateType.MAJOR) {
-      return;
+  _handleFilmsModelEvent(updateType, updatedFilm) {
+    switch (updateType) {
+      case UpdateType.PATCH:
+        this._film = updatedFilm;
+        break;
+      case UpdateType.MINOR:
+        this.init(updatedFilm, { isLoadComments: false });
+        break;
     }
+  }
 
-    this.init(updatedFilm, { loadComments: false });
+  _handleCommentsModelEvent(updateType, updatedPayload) {
+    switch (updateType) {
+      case UpdateType.INIT:
+        this._renderCommentsTitle();
+        this._renderCommentsList();
+        this._renderNewComment();
+        break;
+      case UpdateType.PATCH:
+        this._renderCommentsTitle();
+        if (Array.isArray(updatedPayload)) {
+          this._renderCommentsList();
+        } else {
+          remove(this._commentView.get(updatedPayload));
+          this._commentView.delete(updatedPayload);
+        }
+        break;
+    }
   }
 
   _renderFilmInfo() {
@@ -184,7 +218,8 @@ export default class FilmDetailsPresenter {
   _renderFilmsBottom() {
     this._filmDetailsBottomView = new FilmDetailsBottomView();
     this._renderCommentsContainer();
-    this._renderComments();
+    this._renderCommentsTitle();
+    this._renderCommentsList();
     this._renderNewComment();
     render(this._filmDetailsView, this._filmDetailsBottomView);
   }
@@ -204,23 +239,25 @@ export default class FilmDetailsPresenter {
       return CommentsTitle.ERROR;
     }
 
-    return this._filmComments.length;
+    return this._commentsModel.getAll().length;
   }
 
-  _renderComments() {
+  _renderCommentsTitle() {
     const prevCommentsTitleView = this._commentsTitleView;
-    const prevCommentsListView = this._commentsListView;
-
     this._commentsTitleView =  new CommentsTitleView(this._getCommentsTitle());
-    this._commentsListView = new CommentsListView();
     rerender(this._commentsTitleView, prevCommentsTitleView, this._commentsContainerView);
-    rerender(this._commentsListView, prevCommentsListView, this._commentsContainerView);
+  }
 
-    if (this._isLoading || this._isError) {
+  _renderCommentsList() {
+    if (!this._isSuccess) {
       return;
     }
 
-    this._filmComments.forEach((comment) => {
+    const prevCommentsListView = this._commentsListView;
+    this._commentsListView = new CommentsListView();
+    rerender(this._commentsListView, prevCommentsListView, this._commentsContainerView);
+
+    this._commentsModel.getAll().forEach((comment) => {
       const commentView = new CommentView(comment);
       this._commentView.set(comment.id, commentView);
       commentView.setDeleteButtonClickHandler(this._handleDeleteButtonClick);
@@ -229,8 +266,11 @@ export default class FilmDetailsPresenter {
   }
 
   _renderNewComment() {
-    this._newCommentView = this._newCommentView || new NewCommentView();
-    render(this._commentsContainerView, this._newCommentView);
+    if (this._isSuccess) {
+      this._newCommentView = this._newCommentView || new NewCommentView();
+      this._newCommentView.clearErrorState();
+      render(this._commentsContainerView, this._newCommentView);
+    }
   }
 
   async _renderFilmDetails() {
@@ -244,16 +284,15 @@ export default class FilmDetailsPresenter {
       return;
     }
 
+    let comments = [];
     try {
       this._isError = false;
-      this._filmComments = await this._api.getComments(this._film);
+      comments = await this._api.getComments(this._film.id);
     } catch (error) {
       this._isError = true;
     } finally {
       this._isLoading = false;
-      this._prevScrollTop = this._filmDetailsView.scrollTop;
-      this._renderComments();
-      this._filmDetailsView.scrollTop = this._prevScrollTop;
+      this._commentsModel.setComments(UpdateType.INIT, comments);
     }
   }
 }
